requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database
duckdb>=0.9.0

# Data Collection
httpx>=0.25.0
praw>=7.7.0
tenacity>=8.2.0

# Topic Modeling & NLP
bertopic>=0.15.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
hdbscan>=0.8.33
umap-learn>=0.5.4

# Scheduling
apscheduler>=3.10.0

# API Server
fastapi>=0.104.0
uvicorn>=0.24.0

# Output
jinja2>=3.1.0

# Optional: LLM Summaries
openai>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
